requirements.txt
tensorflow
numpy
pandas
requests
Pillow
scikit-learn
streamlit
